Create Simple.html
The HTML page should be version controlled in GitHub.
The pipeline should be able to run on demand or when a change is made to the HTML page (e.g. CI/CD) 
The pipeline should be able to accept a parameter named "Environment" which will replace the string "{{ Env }}" inside the HTML page with the parameter entered by the user when the pipeline is run.